Replace old simulator by new simulator.
