+++ src/Simulation/Simulators/ToffoliSimulator/Extensions.cs
@@ -5,9 +5,6 @@ using System;
 using System.Linq;
 using System.Collections;
 using System.Collections.Generic;
-using Microsoft.Quantum.Simulation.Common;
-using Microsoft.Quantum.Simulation.Core;
-using Microsoft.Quantum.Simulation.Simulators.Exceptions;
 
 namespace Microsoft.Quantum.Simulation.Simulators
 {
